Use puppeteer to generate unit ratios
