+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "puppeteer": "^1.14.0"
+  }
+}
